BackgroundHelper: use double for scroll, round cap for jobs
- Use double for scroll values to get precise scrolling
- Set Qt::RoundCap for QPen so we do not get artifacrts on line joins

+++ src/graph/view/backgroundhelper.h
@@ -19,11 +19,11 @@ class LineGraphScene;
 class BackgroundHelper
 {
 public:
-    static void drawHourPanel(QPainter *painter, const QRectF& rect, int verticalScroll);
+    static void drawHourPanel(QPainter *painter, const QRectF& rect, double verticalScroll);
 
     static void drawBackgroundHourLines(QPainter *painter, const QRectF& rect);
 
-    static void drawStationHeader(QPainter *painter, LineGraphScene *scene, const QRectF& rect, int horizontalScroll);
+    static void drawStationHeader(QPainter *painter, LineGraphScene *scene, const QRectF& rect, double horizontalScroll);
 
     static void drawStations(QPainter *painter, LineGraphScene *scene, const QRectF& rect);
